dicom archive refactoring
